+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 3.6.0
 - Adds param `hudComponents` in `BonfireWidget`.
+- Adds `queryHud` method in `BonfireGameInterface`
+- Adds `addHud` method in `BonfireGameInterface`
 - UPdate FLame to `1.16.0`
 
 # 3.5.0